+++ FirebaseAuth/Sources/Backend/FIRIdentityToolkitRequest.m
@@ -16,6 +16,8 @@
 
 #import "FirebaseAuth/Sources/Backend/FIRIdentityToolkitRequest.h"
 
+#import "FirebaseAuth/Sources/Public/FirebaseAuth/FIRAuth.h"
+
 NS_ASSUME_NONNULL_BEGIN
 
 static NSString *const kFirebaseAuthAPIURLFormat =
@@ -48,6 +50,14 @@ static NSString *kIdentityPlatformStagingAPIHost =
     _requestConfiguration = requestConfiguration;
     _useIdentityPlatform = NO;
     _useStaging = NO;
+
+    // Automatically set the tenant ID. If the request is initialized before FIRAuth is configured,
+    // set tenant ID to nil.
+    @try {
+      _tenantID = [FIRAuth auth].tenantID;
+    } @catch (NSException *e) {
+      _tenantID = nil;
+    }
   }
   return self;
 }